move most of __init__ to core
